+++ deltafi-core-domain/src/main/java/org/deltafi/core/domain/validation/EnrichFlowPlanValidator.java
@@ -18,6 +18,7 @@
 package org.deltafi.core.domain.validation;
 
 import org.deltafi.core.domain.configuration.ActionConfiguration;
+import org.deltafi.core.domain.configuration.ActionRequiringDomainConfiguration;
 import org.deltafi.core.domain.generated.types.FlowConfigError;
 import org.deltafi.core.domain.services.EnrichFlowPlanService;
 import org.deltafi.core.domain.types.EnrichFlowPlan;
@@ -47,40 +48,42 @@ public class EnrichFlowPlanValidator extends FlowPlanValidator<EnrichFlowPlan> {
         errors.addAll(checkForDuplicatesInOtherPlans(flowPlan));
         errors.addAll(findDuplicatesInFlowPlan(flowPlan));
 
-        return errors;
+        return errors.stream().distinct().collect(Collectors.toList());
     }
 
     List<FlowConfigError> findDuplicatesInFlowPlan(EnrichFlowPlan flowPlan) {
         List<FlowConfigError> errors = new ArrayList<>();
         errors.addAll(findDuplicateActionTypes(flowPlan.getEnrichActions(), flowPlan.getName()));
         errors.addAll(findDuplicateActionTypes(flowPlan.getDomainActions(), flowPlan.getName()));
-        return errors;
+        return errors.stream().distinct().collect(Collectors.toList());
     }
 
     /**
-     * Check the given action list for duplicate action types
+     * Check the given action list for duplicate action types that contain at least one of the same required domains
      * @param actions list of actions to check
      * @param flowPlanName name used in the error message if duplicates exist
      * @return list of errors if duplicate action types are found
      */
-    List<FlowConfigError> findDuplicateActionTypes(List<? extends ActionConfiguration> actions, String flowPlanName) {
-        if (null == actions) {
-            return List.of();
+    List<FlowConfigError> findDuplicateActionTypes(List<? extends ActionRequiringDomainConfiguration> actions, String flowPlanName) {
+        if (null == actions || actions.size() <= 1) {
+            return Collections.emptyList();
         }
 
-        Map<String, List<ActionConfiguration>> duplicatesInflow = actions.stream()
-                .collect(Collectors.groupingBy(ActionConfiguration::getType));
+        List<FlowConfigError> errors = new ArrayList<>();
+        List<? extends ActionRequiringDomainConfiguration> compareActions = new ArrayList<>(actions);
+        ActionRequiringDomainConfiguration action = compareActions.remove(0);
+        while (!compareActions.isEmpty()) {
+            errors.addAll(checkForDuplicatesInOtherPlan(List.of(action), compareActions, flowPlanName, flowPlanName));
+            action = compareActions.remove(0);
+        }
 
-        return duplicatesInflow.entrySet().stream()
-                .filter(entry -> entry.getValue().size() > 1)
-                .map(entry -> duplicateActionError(entry, flowPlanName))
-                .collect(Collectors.toList());
+        return errors.stream().distinct().collect(Collectors.toList());
     }
 
     /**
      * Find any action types that are already configured in an existing
      * flow plan.
-     * @param enrichFlowPlan to search for duplicate enrich action types
+     * @param enrichFlowPlan to search for duplicate enrich action types that contain at least one of the same required domains
      * @return list of errors if new enrich action types already exist in other plans
      */
     List<FlowConfigError> checkForDuplicatesInOtherPlans(EnrichFlowPlan enrichFlowPlan) {
@@ -88,35 +91,38 @@ public class EnrichFlowPlanValidator extends FlowPlanValidator<EnrichFlowPlan> {
                 .filter(otherPlan -> !otherPlan.getName().equals(enrichFlowPlan.getName()))
                 .map(otherPlan -> checkForDuplicatesInOtherPlan(enrichFlowPlan, otherPlan))
                 .flatMap(Collection::stream)
+                .distinct()
                 .collect(Collectors.toList());
     }
 
     private List<FlowConfigError> checkForDuplicatesInOtherPlan(EnrichFlowPlan incomingPlan, EnrichFlowPlan otherPlan) {
         List<FlowConfigError> errors = new ArrayList<>();
-        errors.addAll(checkForDuplicatesInOtherPlan(incomingPlan.getDomainActions(), otherPlan.domainActionTypes(), incomingPlan.getName(), otherPlan.getName()));
-        errors.addAll(checkForDuplicatesInOtherPlan(incomingPlan.getEnrichActions(), otherPlan.enrichActionTypes(), incomingPlan.getName(), otherPlan.getName()));
-        return errors;
+        errors.addAll(checkForDuplicatesInOtherPlan(incomingPlan.getDomainActions(), otherPlan.getDomainActions(), incomingPlan.getName(), otherPlan.getName()));
+        errors.addAll(checkForDuplicatesInOtherPlan(incomingPlan.getEnrichActions(), otherPlan.getEnrichActions(), incomingPlan.getName(), otherPlan.getName()));
+        return errors.stream().distinct().collect(Collectors.toList());
     }
 
-    private List<FlowConfigError> checkForDuplicatesInOtherPlan(List<? extends ActionConfiguration> incomingActions, Set<String> existingActions, String incomingPlanName, String otherPlanName) {
+    private List<FlowConfigError> checkForDuplicatesInOtherPlan(List<? extends ActionRequiringDomainConfiguration> incomingActions, List<? extends ActionRequiringDomainConfiguration> existingActions, String incomingPlanName, String otherPlanName) {
         return incomingActions != null ? findDuplicateActionConfig(incomingActions, existingActions).stream()
                 .map(dupeAction -> duplicateActionError(dupeAction, incomingPlanName, otherPlanName))
-                .collect(Collectors.toList()) : List.of();
+                .distinct()
+                .collect(Collectors.toList()) : Collections.emptyList();
     }
 
-    private List<ActionConfiguration> findDuplicateActionConfig(List<? extends ActionConfiguration> incomingActions, Set<String> existingActionTypes) {
+    private List<ActionRequiringDomainConfiguration> findDuplicateActionConfig(List<? extends ActionRequiringDomainConfiguration> incomingActions, List<? extends ActionRequiringDomainConfiguration> existingActions) {
         return incomingActions.stream()
-                .filter(action -> existingActionTypes.contains(action.getType()))
+                .filter(action -> existingActions.stream()
+                        .anyMatch(existingAction -> existingAction.getType().equals(action.getType())
+                                && existingAction.getRequiresDomains().stream().anyMatch(domain -> action.getRequiresDomains().contains(domain))))
                 .collect(Collectors.toList());
     }
 
-    private FlowConfigError duplicateActionError(Map.Entry<String, List<ActionConfiguration>> entry, String flowName) {
-        String actions = entry.getValue().stream().map(ActionConfiguration::getName).collect(Collectors.joining(", "));
-        return FlowConfigError.newBuilder().configName(flowName).message("The flow contains the same action type: " + entry.getKey() + " in the following actions: " + actions).build();
-    }
-
-    private FlowConfigError duplicateActionError(ActionConfiguration actionConfiguration, String planName, String otherPlan) {
-        return FlowConfigError.newBuilder().configName(planName).message("Action of type: " + actionConfiguration.getType() + " is already configured in the enrich flow plan named: " + otherPlan).build();
+    private FlowConfigError duplicateActionError(ActionRequiringDomainConfiguration actionConfiguration, String planName, String otherPlan) {
+        return FlowConfigError.newBuilder().configName(planName).message("Action of type: " + actionConfiguration.getType() +
+                " is already configured in the enrich flow plan named: " + otherPlan +
+                " with overlapping domain" +
+                (actionConfiguration.getRequiresDomains().size() > 1 ? "s: at least one of " : ": ")
+                + String.join(", ", actionConfiguration.getRequiresDomains())).build();
     }
 
     private boolean blankList(List<? extends ActionConfiguration> actions) {